Face recognition backend
The backend of the Face Recognition Web Application serves as the backbone of the entire system. It manages the business logic, communicates with the PostgreSQL database, and interacts with external services like the Clairfi Face Recognition API. Through the use of Node.js and npm, the backend provides a reliable and efficient foundation for processing user requests, managing data, and ensuring a seamless experience for users interacting with the frontend.